Move IsValidText utility to a special file in common to be used in tests only
