+++ pkg/common/utils.go
@@ -251,41 +251,3 @@ func init() {
 func Tokenize(text string) []string {
 	return re.FindAllString(text, -1)
 }
-
-// IsValidText validates that the given text could be generated from the predefined list of sentences
-// used in tests
-func IsValidText(text string) bool {
-	charsTested := 0
-
-	for charsTested < len(text) {
-		textToCheck := text[charsTested:]
-		found := false
-
-		for _, fakeSentense := range chatCompletionFakeResponses {
-			if len(textToCheck) <= len(fakeSentense) {
-				if strings.HasPrefix(fakeSentense, textToCheck) {
-					found = true
-					charsTested = len(text)
-					break
-				}
-			} else {
-				if strings.HasPrefix(textToCheck, fakeSentense) {
-					charsTested += len(fakeSentense)
-					// during generation sentences are connected by space, skip it
-					// additional space at the end of the string is invalid
-					if text[charsTested] == ' ' && charsTested < len(text)-1 {
-						charsTested += 1
-						found = true
-					}
-					break
-				}
-			}
-		}
-
-		if !found {
-			return false
-		}
-	}
-
-	return true
-}
